Speeding up the checkpoint loader using shared memory for copying (#89)
Previously, the checkpoint loader used ptrace POKE_TEXT API to copy the data from the tracer to tracee, which was too slow because POKE_TEXT involves a system call for copying only 8 bytes of data. This commit changes the copying mechanism by allocating a shared memory buffer for the tracer and tracee and injecting code in the tracee to do the copy with much fewer system calls.

+++ utils/checkpoint/loader/utils.cc
@@ -32,27 +32,40 @@
 #include <stdarg.h>
 #include <sys/personality.h>
 
-void vfatal(const char* fmt, va_list va) {
+void vfatal(const char* fmt, ...) {
+  va_list va;
+  va_start(va, fmt);
+
   fprintf(stderr, "fatal: ");
   vfprintf(stderr, fmt, va);
   fprintf(stderr, "\n");
+  va_end(va);
   exit(1);
 }
 
-void fatal_and_kill_child(pid_t child_pid, const char* fmt, va_list va) {
-  assert(0 != child_pid);
-  kill(child_pid, SIGKILL);
+void fatal_and_kill_child(pid_t child_pid, const char* fmt, ...) {
+  va_list va;
+  va_start(va, fmt);
+
+  if(0 != child_pid) {
+    kill(child_pid, SIGKILL);
+  }
 
   vfatal(fmt, va);
+  va_end(va);
 }
 
 void debug(const char* fmt, ...) {
-  va_list va;
-  va_start(va, fmt);
-  fprintf(stdout, "DEBUG: ");
-  vfprintf(stdout, fmt, va);
-  fprintf(stdout, "\n");
-  va_end(va);
+  if(DEBUG_EN) {
+    va_list va;
+    va_start(va, fmt);
+
+    fprintf(stdout, "DEBUG: ");
+    vfprintf(stdout, fmt, va);
+    fprintf(stdout, "\n");
+    fflush(stdout);
+    va_end(va);
+  }
 }
 
 void assertm(bool p, const char* message) {
@@ -87,8 +100,11 @@ int count_longest_option_length(const struct option long_options[]) {
 
 void turn_aslr_off() {
   const int current_persona = personality(0xffffffff);
-  assert(-1 != current_persona);
-  int ret_val = personality(current_persona | ADDR_NO_RANDOMIZE);
+  if(-1 == current_persona) {
+    vfatal("could not get the current personality");
+  }
+  [[maybe_unused]] int ret_val = personality(current_persona |
+                                             ADDR_NO_RANDOMIZE);
   assert(current_persona == ret_val);
 }
